requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
asyncpg==0.27.0
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.7
python-dotenv==1.0.0
